moved constants to separate file
