+++ src/app/pages/TopicPage/Curation/index.test.jsx
@@ -3,8 +3,8 @@ import React from 'react';
 import fixture from '#data/pidgin/topics/c95y35941vrt.json';
 import mundoFixture from '#data/mundo/topics/c1en6xwmpkvt.json';
 import { render } from '../../../components/react-testing-library-with-providers';
-
-import Curation, { VISUAL_STYLE, VISUAL_PROMINENCE } from '.';
+import { VISUAL_PROMINENCE, VISUAL_STYLE } from '../constants';
+import Curation from '.';
 
 jest.mock('../../../components/ThemeProvider');
