+++ src/app/pages/TopicPage/TopicPage.jsx
@@ -17,7 +17,8 @@ import TopicTitle from './TopicTitle';
 import TopicDescription from './TopicDescription';
 import Pagination from './Pagination';
 import MessageBanner from '../../components/MessageBanner';
-import Curation, { VISUAL_PROMINENCE, VISUAL_STYLE } from './Curation';
+import { VISUAL_PROMINENCE, VISUAL_STYLE } from './constants';
+import Curation from './Curation';
 
 const TopicPage = ({ pageData }) => {
   const { lang, translations } = useContext(ServiceContext);
