+++ src/app/pages/TopicPage/constants.js
@@ -1,0 +1,9 @@
+export const VISUAL_STYLE = {
+  NONE: 'NONE',
+  BANNER: 'BANNER',
+};
+
+export const VISUAL_PROMINENCE = {
+  NORMAL: 'NORMAL',
+  HIGH: 'HIGH',
+};
